requirements.txt
djangorestframework==3.13.0
Pillow==8.3.1
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework-simplejwt==5.3.1
python-dateutil==2.8.2
django-environ==0.11.2
djoser==2.2.0
psycopg2==2.9.9
paypalrestsdk
